Use environment over ref if specified
https://docs.github.com/en/actions/deployment/security-hardening-your-deployments/configuring-openid-connect-in-amazon-web-services#configuring-the-role-and-trust-policy

My deployments have been failing where the repo uses a GitHub
environment.

According to the docs linked above, the sub claim on the JWT that GitHub
passes to AWS appears to be different in the case when an environment is
used, compared to when it is not.

When an environment is not used, the sub claim refers to the ref. When
an environment is used, it looks like the sub claim refers to the
environment.

I'm not sure if multiple sub claims are passed.